+++ src/main/java/com/diecast/diecast_back/repository/EscalaMiniaturaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.diecast.diecast_back.model.EscalaMiniatura;
 
-public interface EscalaMiniaturaRepository extends JpaRepository<EscalaMiniatura, Short> {
+public interface EscalaMiniaturaRepository extends JpaRepository<EscalaMiniatura, Long> {
 
 	Optional<EscalaMiniatura> findByNome(String nome);
